services/web/requirements.txt
aiosmtpd==1.4.4.post2
alembic==1.12.1
atpublic==4.0
attrs==23.1.0
Authlib==1.3.2
Babel==2.13.1
blinker==1.7.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
defusedxml==0.7.1
dnspython==2.4.2
elastic-transport==8.10.0
elasticsearch==8.11.0
email-validator==2.1.0.post1
Flask==3.0.0
flask-babel==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.2
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.1
gunicorn==21.2.0
httpie==3.2.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
oauthlib==2.1.0
packaging==23.2
proto-plus==1.24.0
protobuf==5.28.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.1
PyJWT==2.8.0
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==13.7.0
rq==1.15.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
